requirements/base.txt
aiogram==3.13.1
python-dotenv==1.0.1
groq==0.11.0